feat(injuredfs): impl rpc server
