+++ hook.go
@@ -2,80 +2,20 @@ package main
 
 import (
 	"github.com/ethercflow/hookfs/hookfs"
-	"math/rand"
-	"regexp"
-	"sync"
-	"time"
 )
-
-var (
-	faultMap map[string]*faultContext
-	fml      sync.Mutex
-)
-
-func init() {
-	faultMap = make(map[string]*faultContext)
-}
-
-type faultContext struct {
-	errno  error
-	random bool
-	pct    int
-	path   string
-	delay  time.Duration
-}
-
-func randomErrno() error {
-	return nil
-}
-
-func probab(percentage int) bool {
-	return rand.Intn(99) < percentage
-}
-
-func faultInject(path, method string) error {
-	fml.Lock()
-	fc, ok := faultMap[method]
-	if !ok {
-		fml.Unlock()
-		return nil
-	}
-	fml.Unlock()
-
-	if !probab(fc.pct) {
-		return nil
-	}
-
-	if len(fc.path) > 0 {
-		re, err := regexp.Compile(fc.path)
-		if err != nil || !re.MatchString(path) {
-			return nil
-		}
-	}
-
-	var errno error = nil
-	if fc.errno != nil {
-		errno = fc.errno
-	} else if fc.random {
-		errno = randomErrno()
-	}
-
-	if fc.delay > 0 {
-		time.Sleep(fc.delay)
-	}
-
-	return errno
-}
 
 // implements hookfs.HookContext
 type InjuredHookContext struct {
 }
 
 type InjuredHook struct {
+	addr string
 }
 
 // implements hookfs.HookWithInit
 func (h *InjuredHook) Init() error {
+	StartServer(h.addr)
+	return nil
 }
 
 // if hooked is true, the real open() would not be called
