+++ main.go
@@ -3,24 +3,22 @@ package main
 import (
 	"flag"
 	"fmt"
-	"log"
 	"os"
 	"os/signal"
-	"sync"
 	"syscall"
 	"time"
 
-	"github.com/osrg/hookfs/hookfs"
+	"github.com/ethercflow/hookfs/hookfs"
+	log "github.com/sirupsen/logrus"
 )
 
 var (
-	globalInjuredHook *InjuredHook
-
 	addr       = flag.String("addr", ":65534", "The address to bind to")
 	original   = flag.String("original", "", "ORIGINAL")
 	mountpoint = flag.String("mountpoint", "", "MOUNTPOINT")
 	logLevel   = flag.Int("log-level", 0, fmt.Sprintf("log level (%d..%d)", hookfs.LogLevelMin, hookfs.LogLevelMax))
 )
+
 
 func main() {
 	sc := make(chan os.Signal, 1)
@@ -51,22 +49,10 @@ func main() {
 	}
 
 	hookfs.SetLogLevel(*logLevel)
-	globalInjuredHook = &InjuredHook{
-		read:  make(map[string]time.Duration),
-		rl:    &sync.RWMutex{},
-		fsync: make(map[string]time.Duration),
-		fl:    &sync.RWMutex{},
-	}
-	fs, err := hookfs.NewHookFs(*original, *mountpoint, globalInjuredHook)
+	fs, err := hookfs.NewHookFs(*original, *mountpoint, &InjuredHook{addr: *addr})
 	if err != nil {
 		log.Fatal(err)
 	}
-
-	server := &Server{}
-	go func() {
-		log.Fatal(server.Run(*addr))
-	}()
-
 	if err = fs.Serve(); err != nil {
 		log.Fatal(err)
 	}
